services/logger/pyproject.toml
[project]
name = "logger"
version = "0.1.0"
description = "Logger service that listens to NATS command/response streams and logs to PostgreSQL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "puda-db",
    "puda-python",
    "python-dotenv>=1.2.1",
]

[tool.uv.sources]
puda-db = {workspace=true}
puda-python = {workspace=true}

